requirements.txt
Django==2.2.5
psycopg2-binary==2.8.3
raven==6.10.0
python-dotenv==0.10.3
djangorestframework==3.10.3
coreapi==2.3.3
Markdown==3.1.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-parler==2.0
Pillow==6.1.0
django-ckeditor==5.7.1
django-cors-headers==3.1.0
easy-thumbnails==2.6
django-storages[google]==1.7.2
bleach==3.1.0
django-extensions==2.2.1
requests==2.22.0
